move selective rendering hook for modal to separate file
